requirements.txt
Flask==3.0.2
firebase-admin==6.4.0
openai==1.12.0
python-dotenv==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
requests==2.31.0
python-jose==3.3.0
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.27
beautifulsoup4==4.12.3
google-auth==2.28.1
google-auth-oauthlib==1.2.0
Flask-Session==0.5.0 